services/graph-rag-indexer/pyproject.toml
[project]
name = "daes-graph-rag-ingester"
version = "1.0.0"
description = "Ingester sidecar for DAES GraphRAG: World Bank, UN Comtrade, AIS, Chainlink."
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.115.5",
  "uvicorn[standard]==0.32.1",
  "weaviate-client==4.9.4",
  "httpx==0.27.2",
  "websockets==13.1",
  "pyyaml==6.0.2",
  "prometheus-client==0.21.0",
  "pydantic==2.9.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
